Initial commit pipeline interface
